split up bfs and dfs again
Turns out the diff. between dfs and bfs is more than just the data
structure used to keep track of vertices to visit next (stack vs queue).

The other major changes is where to check whether a vertex has been
visited before. In bfs this can happen before adding adjacent vertices
to the queue; in dfs this must only happen once a vertex is popped off
the stack.

Last not least, the logic when to trigger the pre-/post-visit callbacks
needs to change accordingly as well.